WayPoint/WayPointRenderer: Fix unsigned arithmetic in runway rendering
The runway length was changed from an int to uint16_t in commit
d57a6ef34775b33e5618a71b841bd1443666e5df. This caused underflows
in the runway rendering calculations which relied on negative values.

+++ src/WayPoint/WayPointRenderer.cpp
@@ -555,7 +555,7 @@ WayPointRenderer::DrawLandableSymbol(Canvas &canvas, const RasterPoint &pt,
     fixed len;
     if (Appearance.ScaleRunwayLength && runway.IsLengthDefined())
       len = (radius / fixed_two) +
-        ((runway.GetLength() - 500) / 500) * (radius / fixed_four);
+        (((int) runway.GetLength() - 500) / 500) * (radius / fixed_four);
     else
       len = radius;
     len += fixed_two * scale;
